backend/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
phonenumbers==9.0.0
pillow==11.1.0
propcache==0.3.0
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
stripe==11.5.0
twilio==9.4.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
